Add .ipynb file to create video DB
Created and added a new file create_video_db.ipynb where the process to create and populate a new db starting from the data from a JSON file, is described. The code is clearly showing the steps and the function used in this process. Many functions are imported from the mongo_utils.py file.