Add the file create_transcript_db
Created and added the file in the object. a .ipynb file. There is the description step by step of how to get the general information of the Lex Podcast video, retrieve the transcript - if possible - and re structure all the information so that each video along with relevant info and the transcript can be uploaded to a new MongoDB collection. There is the option to create also a local JSON repository.