First RAG version
First creation of a simple .py file where 3 classes where created to handle the RAG system. the first class deals with the connection with MongoDB, the second deals with retrieving information for the selected video and the 3rd deals with the RAG itself. This is the first working version (simple and naive). Notice that for now we are using as default the MongoDB and that we are storing and keeping the vector store locally in a folder named chroma_db.  Additionally requirements.txt and .gitignore were updated.