Add new file rag_video_streamlit.py
Added a new file which is going to use the classes defined in the rag_single_video.py to generate a user friendly interface and a chatbot able to load interviews either via mongoDB or JSON storage and let the user chat with the bot. This is the first working version.