Update LLM
Update options to use also Anthropic as LLM for generation. Unfortunately it appears that Anthropic has not created its own embedding models like OpenAI has. Hence for Anthropic we are going to still use OpenAI embeddings. Modified also the hard reset part making such that every past message is deleted and cancelled from the app and a nicer apologize message is printed. The limit is set to 50 for now.